Seperate sum encoding
